+++ src/dummy_graph.ts
@@ -1,241 +1,143 @@
-import { IGraph, EGraphNodeType } from "./types";
+import { QueryResponse } from "./types";
 
-const graph: IGraph = {
-    nodes: [
-        {
-            name: "gateway",
-            type: EGraphNodeType.INTERNAL,
-            metrics: {
-                rate: 550,
-                error_rate: 25,
-                response_time: 38
-            }
-        },
-        {
-            name: "customer-service-1",
-            type: EGraphNodeType.INTERNAL,
-            metrics: {
-                rate: 125,
-                error_rate: 100,
-                response_time: 24
-            }
-        },
-        {
-            name: "customer-service-2",
-            type: EGraphNodeType.INTERNAL,
-            metrics: {
-                rate: 225,
-                error_rate: 5,
-                response_time: 25
-            }
-        },
-        {
-            name: "payment-service",
-            type: EGraphNodeType.INTERNAL,
-            metrics: {
-                rate: 300,
-                error_rate: 30,
-                response_time: 18
-            }
-        },
-        {
-            name: "api.finance.com",
-            type: EGraphNodeType.EXTERNAL,
-            external_type: "http",
-            metrics: {
-                rate: 750,
-                error_rate: 0,
-                response_time: 134
-            }
-        },
-        {
-            name: "https://lb-customer.local",
-            type: EGraphNodeType.EXTERNAL,
-            external_type: 'balancer',
-            metrics: {
-                rate: 350
-            }
-        },
-        {
-            name: "inventory-service",
-            type: EGraphNodeType.INTERNAL,
-            metrics: {
-                rate: 200,
-                error_rate: 0,
-                response_time: 12
-            }
-        },
-        {
-            name: "jdbc:oracle:thin:@sample-oracle-db:1521",
-            type: EGraphNodeType.EXTERNAL,
-            external_type: 'jdbc',
-            metrics: {
-                rate: 1200,
-                response_time: 2
-            }
-        },
-        {
-            name: "api.productsupplier.com",
-            type: EGraphNodeType.EXTERNAL,
-            external_type: "http",
-            metrics: {
-                rate: 600,
-                response_time: 83
-            }
-        },
-        {
-            name: "tcp://localhost:61616",
-            type: EGraphNodeType.EXTERNAL,
-            external_type: "jms",
-            metrics: {
-                rate: 300,
-                response_time: 7
-            }
-        },
-        {
-            name: "reporting-service",
-            type: EGraphNodeType.INTERNAL,
-            metrics: {
-                rate: 160,
-                response_time: 18
-            }
-        },
-        {
-            name: "controlling-service",
-            type: EGraphNodeType.INTERNAL,
-            metrics: {
-                rate: 300,
-                response_time: 45
-            }
-        }
-    ],
-    edges: [
-        {
-            source: "tcp://localhost:61616",
-            target: "reporting-service",
-            metrics: {
-                rate: 300,
-                response_time: 20
-            }
-        },
-        {
-            source: "controlling-service",
-            target: "reporting-service",
-            metrics: {
-                rate: 10,
-                response_time: 12
-            }
-        },
-        {
-            source: "controlling-service",
-            target: "jdbc:oracle:thin:@sample-oracle-db:1521",
-            metrics: {
-                rate: 300,
-                response_time: 2
-            }
-        },
-        {
-            source: "tcp://localhost:61616",
-            target: "controlling-service",
-            metrics: {
-                rate: 300,
-                response_time: 45
-            }
-        },
-        {
-            source: "payment-service",
-            target: "tcp://localhost:61616",
-            metrics: {
-                rate: 300,
-                response_time: 7
-            }
-        },
-        {
-            source: "gateway",
-            target: "https://lb-customer.local",
-            metrics: {
-                rate: 350,
-                response_time: 35,
-                error_rate: 30
-            }
-        },
-        {
-            source: "gateway",
-            target: "inventory-service",
-            metrics: {
-                rate: 200,
-                response_time: 15
-            }
-        },
-        {
-            source: "inventory-service",
-            target: "jdbc:oracle:thin:@sample-oracle-db:1521",
-            metrics: {
-                rate: 400,
-                response_time: 1
-            }
-        },
-        {
-            source: "payment-service",
-            target: "jdbc:oracle:thin:@sample-oracle-db:1521",
-            metrics: {
-                rate: 500,
-                response_time: 2
-            }
-        },
-        {
-            source: "https://lb-customer.local",
-            target: "customer-service-1",
-            metrics: {
-                rate: 125,
-                error_rate: 25,
-                response_time: 45
-            }
-        },
-        {
-            source: "https://lb-customer.local",
-            target: "customer-service-2",
-            metrics: {
-                rate: 225,
-                error_rate: 5,
-                response_time: 25
-            }
-        },
-        {
-            source: "customer-service-1",
-            target: "payment-service",
-            metrics: {
-                rate: 125,
-                error_rate: 100,
-                response_time: 23
-            }
-        },
-        {
-            source: "customer-service-2",
-            target: "payment-service",
-            metrics: {
-                rate: 225,
-                error_rate: 5,
-                response_time: 22
-            }
-        },
-        {
-            source: "payment-service",
-            target: "api.finance.com",
-            metrics: {
-                rate: 750,
-                response_time: 134
-            }
-        },
-        {
-            source: "inventory-service",
-            target: "api.productsupplier.com",
-            metrics: {
-                rate: 600,
-                response_time: 83
-            }
-        }
-    ]
-};
+const rowData: QueryResponse[] = [
+    {
+        columns: [
+            { type: "time", text: "Time" },
+            { text: "origin_external" },
+            { text: "origin_service" },
+            { text: "protocol" },
+            { text: "service" },
+            { text: "in_count" }
+        ],
+        refId: undefined,
+        meta: undefined,
+        rows: [
+            [0, "", "", "http", "api-gateway", 508],
+            [0, "", "", "http", "config-server", 0],
+            [0, "", "", "http", "discovery-server", 0],
+            [0, "", "api-gateway", "http", "api-gateway", 100],
+            [0, "", "api-gateway", "http", "customers-service", 347],
+            [0, "", "api-gateway", "http", "discovery-server", 20],
+            [0, "", "api-gateway", "http", "vets-service", 63],
+            [0, "", "api-gateway", "http", "visits-service", 100],
+            [0, "", "customers-service", "http", "discovery-server", 20],
+            [0, "", "vets-service", "http", "discovery-server", 20],
+            [0, "", "visits-service", "http", "discovery-server", 20],
+            [0, "MyService", "vets-service", "http", "visits-service", 300]
+        ]
+    },
 
-export default graph;
+    {
+        columns: [
+            { type: "time", text: "Time" },
+            { text: "protocol" },
+            { text: "service" },
+            { text: "target_external" },
+            { text: "target_service" },
+            { text: "out_count" }
+        ],
+        refId: undefined,
+        meta: undefined,
+        rows: [
+            [0, "http", "api-gateway", "", "api-gateway", 100],
+            [0, "http", "api-gateway", "", "customers-service", 347],
+            [0, "http", "api-gateway", "", "discovery-server", 20],
+            [0, "http", "api-gateway", "", "vets-service", 62],
+            [0, "http", "api-gateway", "", "visits-service", 100],
+            [0, "http", "api-gateway", "7a8dce897616:8080", "", 0],
+            [0, "http", "config-server", "github.com", "", 0],
+            [0, "http", "customers-service", "", "discovery-server", 20],
+            [0, "http", "vets-service", "", "discovery-server", 20],
+            [0, "http", "vets-service", "MyService", "visits-service", 300],
+            [0, "http", "visits-service", "", "discovery-server", 20],
+            [0, "jdbc", "customers-service", "jdbc:hsqldb:mem:testdb", "", 1847],
+            [0, "jdbc", "vets-service", "jdbc:hsqldb:mem:testdb", "", 441],
+            [0, "jdbc", "visits-service", "jdbc:hsqldb:mem:testdb", "", 100]
+        ]
+    },
+
+    {
+        columns: [
+            { type: "time", text: "Time" },
+            { text: "origin_service" },
+            { text: "protocol" },
+            { text: "service" },
+            { text: "target_external" },
+            { text: "in_timesum" }
+        ],
+        refId: undefined,
+        meta: undefined,
+        rows: [
+            [0, "", "http", "api-gateway", "", 4514.008427999986],
+            [0, "", "http", "config-server", "", 0],
+            [0, "", "http", "discovery-server", "", 0],
+            [0, "api-gateway", "http", "api-gateway", "", 1511.9842349999872],
+            [0, "api-gateway", "http", "customers-service", "", 819.3634589999965],
+            [0, "api-gateway", "http", "discovery-server", "", 21.881731999999943],
+            [0, "api-gateway", "http", "vets-service", "", 281.0465210000002],
+            [0, "api-gateway", "http", "visits-service", "", 325.85070300000007],
+            [0, "customers-service", "http", "discovery-server", "", 21.53124399999996],
+            [0, "vets-service", "http", "discovery-server", "", 21.40604300000001],
+            [0, "visits-service", "http", "discovery-server", "", 20.813048000000038]
+        ]
+    },
+
+    {
+        columns: [
+            { type: "time", text: "Time" },
+            { text: "protocol" },
+            { text: "service" },
+            { text: "target_external" },
+            { text: "target_service" },
+            { text: "out_timesum" }
+
+        ],
+        refId: undefined,
+        meta: undefined,
+        rows: [
+            [0, "http", "api-gateway", "", "api-gateway", 1700.468872999987],
+            [0, "http", "api-gateway", "", "customers-service", 1481.533606999972],
+            [0, "http", "api-gateway", "", "discovery-server", 540.746261],
+            [0, "http", "api-gateway", "", "vets-service", 501.65547400000014],
+            [0, "http", "api-gateway", "", "visits-service", 394.81158100000175],
+            [0, "http", "api-gateway", "7a8dce897616:8080", "", 0],
+            [0, "http", "config-server", "github.com", "", 0],
+            [0, "http", "customers-service", "", "discovery-server", 84.59527999999978],
+            [0, "http", "vets-service", "", "discovery-server", 381.87400800000023],
+            [0, "http", "visits-service", "", "discovery-server", 225.65933600000017],
+            [0, "jdbc", "customers-service", "jdbc:hsqldb:mem:testdb", "", 35.9093940000007],
+            [0, "jdbc", "vets-service", "jdbc:hsqldb:mem:testdb", "", 13.000189000000091],
+            [0, "jdbc", "visits-service", "jdbc:hsqldb:mem:testdb", "", 12.258137999999946]
+        ]
+    },
+
+    {
+        columns: [
+            { type: "time", text: "Time" },
+            { text: "origin_service" },
+            { text: "protocol" },
+            { text: "service" },
+            { text: "target_external" },
+            { text: "error" }
+        ],
+        refId: undefined,
+        meta: undefined,
+        rows: [
+            [0, "", "http", "api-gateway", "", 254.5],
+            [0, "", "http", "config-server", "", 0],
+            [0, "", "http", "discovery-server", "", 0],
+            [0, "api-gateway", "http", "api-gateway", "", 49.5],
+            [0, "api-gateway", "http", "customers-service", "", 173.5],
+            [0, "api-gateway", "http", "discovery-server", "", 10],
+            [0, "api-gateway", "http", "vets-service", "", 31.5],
+            [0, "api-gateway", "http", "visits-service", "", 49.5],
+            [0, "customers-service", "http", "discovery-server", "", 10],
+            [0, "vets-service", "http", "discovery-server", "", 10],
+            [0, "vets-service", "http", "visits-service", "", 150],
+            [0, "visits-service", "http", "discovery-server", "", 10],
+        ]
+    }
+]
+
+export default rowData;
